YAMNet/requirements.txt
numpy
tensorflow>=2.5
resampy
pysoundfile
